:recycle: split measure.h and measure.cc
